+++ src/deluge/gui/menu_item/param.cpp
@@ -17,12 +17,17 @@
 
 #include "param.h"
 #include "gui/l10n/l10n.h"
+#include "gui/ui/sound_editor.h"
+#include "gui/views/automation_view.h"
 #include "hid/buttons.h"
 #include "hid/display/display.h"
 #include "model/action/action.h"
 #include "model/action/action_logger.h"
+#include "model/clip/clip.h"
 #include "model/model_stack.h"
+#include "model/song/song.h"
 #include "modulation/automation/auto_param.h"
+#include "modulation/params/param_set.h"
 
 namespace deluge::gui::menu_item {
 
@@ -42,4 +47,58 @@ MenuItem* Param::selectButtonPress() {
 	display->displayPopup(l10n::get(l10n::String::STRING_FOR_AUTOMATION_DELETED));
 	return (MenuItem*)0xFFFFFFFF; // Don't navigate away
 }
+
+ActionResult Param::buttonAction(deluge::hid::Button b, bool on) {
+	using namespace deluge::hid::button;
+
+	bool clipMinder = rootUIIsClipMinderScreen();
+	bool arrangerView = !clipMinder && (currentSong->lastClipInstanceEnteredStartPos != -1);
+	RootUI* rootUI = getRootUI();
+
+	// Clip or Song button
+	// Used to enter automation view from sound editor
+	if ((b == CLIP_VIEW && clipMinder) || (b == SESSION_VIEW && arrangerView)) {
+		if (on) {
+			if (rootUI != &automationView) {
+				selectAutomationViewParameter(clipMinder);
+				swapOutRootUILowLevel(&automationView);
+				automationView.openedInBackground();
+			}
+			soundEditor.exitCompletely();
+		}
+		return ActionResult::DEALT_WITH;
+	}
+	// Select encoder button, used to change current parameter selection in automation view
+	// if you are already in automation view and entered an automatable parameter menu
+	else if (b == SELECT_ENC && (clipMinder || arrangerView)) {
+		if (on) {
+			if (rootUI == &automationView) {
+				selectAutomationViewParameter(clipMinder);
+				uiNeedsRendering(&automationView);
+			}
+		}
+		return ActionResult::DEALT_WITH;
+	}
+	return ActionResult::NOT_DEALT_WITH;
+}
+
+void Param::selectAutomationViewParameter(bool clipMinder) {
+	char modelStackMemory[MODEL_STACK_MAX_SIZE];
+	ModelStackWithAutoParam* modelStack = getModelStack(modelStackMemory);
+
+	int32_t p = getP();
+	modulation::params::Kind kind = modelStack->paramCollection->getParamKind();
+
+	if (clipMinder) {
+		Clip* clip = getCurrentClip();
+		clip->lastSelectedParamID = p;
+		clip->lastSelectedParamKind = kind;
+	}
+	else {
+		currentSong->lastSelectedParamID = p;
+		currentSong->lastSelectedParamKind = kind;
+		automationView.onArrangerView = true;
+	}
+}
+
 } // namespace deluge::gui::menu_item