+++ src/deluge/processing/sound/sound.h
@@ -294,6 +294,4 @@ private:
 	ModelStackWithAutoParam* getParamFromModEncoderDeeper(int32_t whichModEncoder,
 	                                                      ModelStackWithThreeMainThings* modelStack,
 	                                                      bool allowCreation = true);
-
-	void displaySidechainAndReverbSettings(bool on);
 };