+++ src/deluge/util/functions.cpp
@@ -933,31 +933,6 @@ FilterType stringToFilterType(char const* string) {
 	}
 }
 
-char const* filterRouteToString(FilterRoute route) {
-	switch (route) {
-	case FilterRoute::LOW_TO_HIGH:
-		return "L2H";
-
-	case FilterRoute::PARALLEL:
-		return "PARA";
-
-	default:
-		return "H2L";
-	}
-}
-
-FilterRoute stringToFilterRoute(char const* string) {
-	if (!strcmp(string, "L2H")) {
-		return FilterRoute::LOW_TO_HIGH;
-	}
-	else if (!strcmp(string, "PARA")) {
-		return FilterRoute::PARALLEL;
-	}
-	else {
-		return FilterRoute::HIGH_TO_LOW;
-	}
-}
-
 char const* arpModeToString(ArpMode mode) {
 	switch (mode) {
 	case ArpMode::UP:
@@ -992,49 +967,6 @@ ArpMode stringToArpMode(char const* string) {
 	}
 	else {
 		return ArpMode::OFF;
-	}
-}
-// converts lpf/hpf mode to string for saving
-char const* lpfTypeToString(FilterMode lpfType) {
-	switch (lpfType) {
-	case FilterMode::TRANSISTOR_12DB:
-		return "12dB";
-
-	case FilterMode::TRANSISTOR_24DB_DRIVE:
-		return "24dBDrive";
-	case FilterMode::SVF_BAND:
-		return "SVF_Band";
-	case FilterMode::HPLADDER:
-		return "HPLadder";
-	case FilterMode::SVF_NOTCH:
-		return "SVF_Notch";
-	default:
-		return "24dB";
-	}
-}
-
-FilterMode stringToLPFType(char const* string) {
-	if (!strcmp(string, "24dB")) {
-		return FilterMode::TRANSISTOR_24DB;
-	}
-	else if (!strcmp(string, "24dBDrive")) {
-		return FilterMode::TRANSISTOR_24DB_DRIVE;
-	}
-	else if (!strcmp(string, "SVF_Band")) {
-		return FilterMode::SVF_BAND;
-	}
-	else if (!strcmp(string, "SVF")) {
-		// for compatibility with community pre release
-		return FilterMode::SVF_BAND;
-	}
-	else if (!strcmp(string, "HPLadder")) {
-		return FilterMode::HPLADDER;
-	}
-	else if (!strcmp(string, "SVF_Notch")) {
-		return FilterMode::SVF_NOTCH;
-	}
-	else {
-		return FilterMode::TRANSISTOR_12DB;
 	}
 }